Move parsing related files from sympy/core to sympy/parsing
This is sympify.py, ast_parser.py, and ast_parser_python24.py.  The main
reason for doing this is that ast_parser_python24.py requires the Python
standard library symbol module, which could not be imported directly in
the file because of the sympy/core/symbol.py file.  Thus, we had some
chicanery going on where we imported it as stdlib_symbol in
sympy/__init__.py, which was then being put into the global namespace
with every "from sympy import *".  The easiest way to fix this was to
move the file out of sympy/core/, and since those other files are
similar, they too were moved out.

See
http://groups.google.com/group/sympy/browse_thread/thread/f471ea9a8d3ef041
for more information.

This additionally adds SympifyError to the global namespace.
