+++ sympy/__init__.py
@@ -18,8 +18,8 @@ def __sympy_debug():
     return eval(os.getenv('SYMPY_DEBUG', 'False'))
 SYMPY_DEBUG = __sympy_debug()
 
-import symbol as stdlib_symbol
 from sympy.core import *
+from parsing import *
 from assumptions import *
 from polys import *
 from series import *
@@ -47,4 +47,3 @@ evalf._create_evalf_table()
 
 # This is slow to import:
 #import abc
-
