+++ sympy/functions/elementary/piecewise.py
@@ -2,8 +2,9 @@ from sympy.core.basic import Basic, S
 from sympy.core.function import Function, diff
 from sympy.core.numbers import Number
 from sympy.core.relational import Relational
-from sympy.core.sympify import sympify
 from sympy.core.sets import Interval, Set
+
+from sympy.parsing.sympify import sympify
 
 class ExprCondPair(Function):
     """Represents an expression, condition pair."""
