+++ sympy/functions/special/gamma_functions.py
@@ -1,8 +1,9 @@
-from sympy.core import Add, S, C, sympify
+from sympy.core import Add, S, C
 from sympy.core.function import Function
 from zeta_functions import zeta
 from sympy.functions.elementary.exponential import log
 from sympy.functions.elementary.miscellaneous import sqrt
+from sympy.parsing.sympify import sympify
 
 ###############################################################################
 ############################ COMPLETE GAMMA FUNCTION ##########################
