+++ sympy/matrices/matrices.py
@@ -1,6 +1,6 @@
 import warnings
 from sympy import Basic, Symbol, Integer
-from sympy.core.sympify import sympify, converter, SympifyError
+from sympy.parsing.sympify import sympify, converter, SympifyError
 
 from sympy.core.basic import S
 from sympy.polys import Poly, roots, cancel
