+++ sympy/parsing/__init__.py
@@ -1,2 +1,4 @@
 """Used for translating a string into a sympy expression.
 """
+
+from sympify import sympify, SympifyError
