+++ sympy/parsing/ast_parser_python24.py
@@ -4,9 +4,10 @@ from compiler.transformer import Transformer
 from compiler.ast import CallFunc, Name, Const
 from compiler.pycodegen import ExpressionCodeGenerator
 import re
+import symbol # This is Python's stdlib symbol, not SymPy's symbol
 
-from basic import Basic
-from symbol import Symbol
+from sympy.core.basic import Basic
+from sympy.core.symbol import Symbol
 
 _is_integer = re.compile(r'\A\d+(l|L)?\Z').match
 
@@ -52,8 +53,7 @@ class SymPyTransformer(Transformer):
 
     def lambdef(self, nodelist):
         #this is python stdlib symbol, not SymPy symbol:
-        from sympy import stdlib_symbol
-        if nodelist[2][0] == stdlib_symbol.varargslist:
+        if nodelist[2][0] == symbol.varargslist:
             names, defaults, flags = self.com_arglist(nodelist[2][1:])
         else:
             names = defaults = ()
