+++ sympy/parsing/tests/test_sympify.py
@@ -1,7 +1,7 @@
 from sympy import Symbol, exp, Integer, Real, sin, cos, log, Poly, Lambda, \
         Function, I, S, sqrt,  raises, srepr, Rational
 from sympy.abc import x, y
-from sympy.core.sympify import sympify, _sympify, SympifyError
+from sympy.parsing.sympify import sympify, _sympify, SympifyError
 from sympy.core.decorators import _sympifyit
 
 def test_439():
