+++ sympy/polys/polyroots.py
@@ -2,11 +2,11 @@ from sympy.core.symbol import Symbol
 from sympy.core.add import Add
 from sympy.core.mul import Mul
 from sympy.core import Expr, S
-from sympy.core.sympify import sympify
 from sympy.core.numbers import Rational
 
 from sympy.ntheory import divisors
 from sympy.functions import exp, sqrt
+from sympy.parsing.sympify import sympify
 
 from sympy.polys import (
     Poly, cancel, PolynomialError, GeneratorsNeeded,
