+++ sympy/simplify/rewrite.py
@@ -1,8 +1,9 @@
 """Partial fraction decomposition algorithms. """
 
-from sympy.core import sympify, Symbol, Function, Lambda
+from sympy.core import Symbol, Function, Lambda
 from sympy.polys import Poly, RootSum, div, cancel
 from sympy.utilities import threaded
+from sympy.parsing.sympify import sympify
 
 @threaded()
 def apart(f, z=None, **args):
