+++ sympy/simplify/sqrtdenest.py
@@ -1,6 +1,7 @@
 from sympy.utilities import all, any
 from sympy.functions import sqrt, sign
-from sympy.core import Pow, S, Number, Wild, Rational, sympify
+from sympy.core import Pow, S, Number, Wild, Rational
+from sympy.parsing.sympify import sympify
 
 def sqrtdenest (expr):
     """
