+++ sympy/solvers/recurr.py
@@ -52,8 +52,8 @@ from sympy.core.symbol import Symbol, Wild
 from sympy.core.relational import Equality
 from sympy.core.add import Add
 from sympy.core.mul import Mul
-from sympy.core import sympify
 
+from sympy.parsing.sympify import sympify
 from sympy.simplify import simplify, hypersimp, hypersimilar
 from sympy.solvers import solve, solve_undetermined_coeffs
 from sympy.polys import Poly, exquo, gcd, lcm, roots, resultant
