+++ sympy/tensor/indexed.py
@@ -102,7 +102,9 @@
 #      - Idx with stepsize != 1
 #      - Idx with step determined by function call
 
-from sympy.core import Expr, Basic, Tuple, Symbol, Integer, sympify, S
+from sympy.core import Expr, Basic, Tuple, Symbol, Integer, S
+
+from sympy.parsing.sympify import sympify
 
 class IndexException(Exception):
     pass
